Add support for request button actions via FIFO pipe
Following the shell experiment this creates a safer way of interacting with the caller script. This commit introduce a new 'request' button action type in InspectConfig, allowing script request interaction through a configurable FIFO pipe. PresetCommonHelpers allows now to handle 'request' actions and write to the pipe if specified, and propagates requestId in Preset6 layout. This enables external scripts to react to UI button events in near real time.